Revert SchedulerExecutorPoolImpl
