+++ litecommands-annotations/test/dev/rollczi/litecommands/annotations/async/AsyncArgumentsInitializeTest.java
@@ -3,7 +3,6 @@ package dev.rollczi.litecommands.annotations.async;
 import dev.rollczi.litecommands.argument.Argument;
 import dev.rollczi.litecommands.programmatic.LiteCommand;
 import static dev.rollczi.litecommands.programmatic.LiteProgrammatic.async;
-import dev.rollczi.litecommands.scheduler.SchedulerExecutorPoolImpl;
 import dev.rollczi.litecommands.unit.AssertExecute;
 import dev.rollczi.litecommands.unit.LiteCommandsTestFactory;
 import dev.rollczi.litecommands.unit.TestPlatform;
