+++ litecommands-core/test/dev/rollczi/litecommands/argument/resolver/completeable/CompletableFutureArgumentTest.java
@@ -4,7 +4,6 @@ import dev.rollczi.litecommands.annotations.argument.Arg;
 import dev.rollczi.litecommands.annotations.command.Command;
 import dev.rollczi.litecommands.annotations.execute.Execute;
 import dev.rollczi.litecommands.reflect.type.TypeToken;
-import dev.rollczi.litecommands.scheduler.SchedulerExecutorPoolImpl;
 import dev.rollczi.litecommands.unit.annotations.LiteTestSpec;
 import dev.rollczi.litecommands.unit.blocking.Blocking;
 import dev.rollczi.litecommands.unit.blocking.BlockingArgument;
