+++ litecommands-framework/src/dev/rollczi/litecommands/LiteCommandsBaseBuilder.java
@@ -45,7 +45,6 @@ import dev.rollczi.litecommands.reflect.type.TypeRange;
 import dev.rollczi.litecommands.requirement.RequirementMatchService;
 import dev.rollczi.litecommands.scheduler.Scheduler;
 import dev.rollczi.litecommands.scheduler.SchedulerExecutorPoolBuilder;
-import dev.rollczi.litecommands.scheduler.SchedulerExecutorPoolImpl;
 import dev.rollczi.litecommands.scheduler.SchedulerReference;
 import dev.rollczi.litecommands.schematic.SchematicFastFormat;
 import dev.rollczi.litecommands.schematic.SchematicFastGenerator;
