+++ litecommands-minestom/src/dev/rollczi/litecommands/minestom/LiteMinestomFactory.java
@@ -8,7 +8,6 @@ import dev.rollczi.litecommands.minestom.argument.InstanceArgument;
 import dev.rollczi.litecommands.minestom.argument.PlayerArgument;
 import dev.rollczi.litecommands.minestom.context.InstanceContext;
 import dev.rollczi.litecommands.minestom.context.PlayerContext;
-import dev.rollczi.litecommands.scheduler.SchedulerExecutorPoolImpl;
 import net.minestom.server.MinecraftServer;
 import net.minestom.server.command.CommandManager;
 import net.minestom.server.command.CommandSender;
